Make unerInout a separate method in Computer, Date an inner class
